+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD (unreleased)
 
+- "Unmatched end" and "missing end" not generate different error text instructions (https://github.com/zombocom/syntax_search/pull/10)
+
 ## 0.1.1
 
 - Fire search on both unexpected end-of-input and unexected end (https://github.com/zombocom/syntax_search/pull/8)